Rename makeADecisionPage to makeADecisionTask
